Move global functions to corresponding modules
